% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{add.alpha}
\alias{add.alpha}
\title{Add.alpha}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{col}{A color.}
}
\value{
the transparent version of the colour
}
\description{
Converts colours to transparent.
}
\examples{
add.alpha("blue")
}

