% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{bi_endpoints_sim}
\alias{bi_endpoints_sim}
\title{bi_endpoints_sim}
\usage{
bi_endpoints_sim(batch, N)
}
\description{
Simulates bivariate endpoints by adding the distribution of the C and N discrimination
factors to the bivariate normal distribution of the food source generated by the
make_mvnorm function.
}

