% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\docType{data}
\name{d13C}
\alias{d13C}
\title{d13C}
\format{\preformatted{  expression(paste(delta^\{     13 \}, "C (‰)"))
}}
\usage{
d13C
}
\description{
Convenient text for plotting.
}
\keyword{datasets}

