% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\docType{data}
\name{d15N}
\alias{d15N}
\title{d15N}
\format{\preformatted{  expression(paste(delta^\{     15 \}, "N (‰)"))
}}
\usage{
d15N
}
\description{
Convenient text for plotting.
}
\keyword{datasets}

