% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{full_plot}
\alias{full_plot}
\title{This function combines all of the above plots - note that you can select which parameter table
you'd like to use here, for easy comparison}
\usage{
full_plot(modelno, params, s)
}
\description{
This function combines all of the above plots - note that you can select which parameter table
you'd like to use here, for easy comparison
}

