% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{generate_mvnorm}
\alias{generate_mvnorm}
\title{generate_mvnorm}
\usage{
generate_mvnorm(batch, N)
}
\description{
Generates a bivariate normal distribution of the dietary group supplied
(batch) using the mvnorm function
}

