% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{make_endpoints}
\alias{make_endpoints}
\title{Make endpoints}
\usage{
make_endpoints(s, params)
}
\description{
Takes isotope values of the selected groups and divides them by type (Cereal, Pulse or Animal)
and addes the appropriate discrimination factor from the params table, and pools the two standard
deviations. It also assigns the appropriate digestable C and N proportions to each type of mixing
group based on the params table.
}

