% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{make_tint}
\alias{make_tint}
\title{make_tint
Automatically makes tinted shade for plotting. Returns RGB colour.}
\usage{
make_tint(col, tint_factor)
}
\description{
make_tint
Automatically makes tinted shade for plotting. Returns RGB colour.
}

