% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{make_xyvals}
\alias{make_xyvals}
\title{Calculates the hypothetical d13C and d15N ratios for the selected set of proportions (model)
The proportions from the model are modified by the C/N ratio of the digestible C and N.
From the endpoints calculated by the endpoints() function 10,000 normally distributed values
are selected, for each of the endpoint groups. Each of these 10,000 ranomly selected values
are averaged, weighted by the digestible C/N ratio-corrected proportions from the selected model
The distribution of these 10,000 weighted mixtures is reported for both x and y.}
\usage{
make_xyvals(e, model)
}
\description{
Calculates the hypothetical d13C and d15N ratios for the selected set of proportions (model)
The proportions from the model are modified by the C/N ratio of the digestible C and N.
From the endpoints calculated by the endpoints() function 10,000 normally distributed values
are selected, for each of the endpoint groups. Each of these 10,000 ranomly selected values
are averaged, weighted by the digestible C/N ratio-corrected proportions from the selected model
The distribution of these 10,000 weighted mixtures is reported for both x and y.
}

