% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params3}
\alias{params3}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Type}{price, in US dollars}
  \item{D13C}{weight of the diamond, in carats}
  \item{D13Csd}{weight of the diamond, in carats}
  \item{D15N}{weight of the diamond, in carats}
  \item{D15Nsd}{weight of the diamond, in carats}
  \item{pcC}{weight of the diamond, in carats}
  \item{pcN}{weight of the diamond, in carats}
  \item{DigestC}{weight of the diamond, in carats}
  \item{DigestN}{weight of the diamond, in carats}
}}
\usage{
params3
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

