% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{plotmyxy}
\alias{plotmyxy}
\title{plotmyxy}
\usage{
plotmyxy(df, bigD = FALSE, dair = NULL, ...)
}
\description{
Makes a standard isotope plot using the standard plotting charactes from the standard.species()
function. Can take all the normal graphical parameters for further customization. It can also
use big D13C or little d13C. The default (bigD=FALSE) is for little d13C. If bigD is selected, the
D13C column must be included in the original data.frame.
}

