% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{A dataset containing the prices and other attributes of almost 54,000
diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{Value}{price, in US dollars}
  \item{Lookup}{weight of the diamond, in carats}
}}
\usage{
species
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

