% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{ID}{price, in US dollars}
  \item{species}{weight of the diamond, in carats}
  \item{normd15N}{weight of the diamond, in carats}
  \item{normd13C}{weight of the diamond, in carats}
}}
\usage{
test_df
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

