% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{model_plot}
\alias{model_plot}
\title{model_plot function - version 2 using kde instead of histogram, based on multivariate distribution}
\usage{
model_plot(s, model.no, ...)
}
\description{
model_plot function - version 2 using kde instead of histogram, based on multivariate distribution
}

