% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_params}
\alias{plot_params}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{English}{price, in US dollars}
  \item{Latin}{weight of the diamond, in carats}
  \item{Alt_name}{weight of the diamond, in carats}
  \item{col.list}{weight of the diamond, in carats}
  \item{bg.list}{weight of the diamond, in carats}
  \item{pch.list}{weight of the diamond, in carats}
  \item{Order}{weight of the diamond, in carats}
  \item{Group}{weight of the diamond, in carats}
  \item{Gp.col}{weight of the diamond, in carats}
  \item{Gp.bg}{weight of the diamond, in carats}
  \item{Type}{weight of the diamond, in carats}
}}
\usage{
plot_params
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

