% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isotopevis.R
\name{standard_species}
\alias{standard_species}
\title{Standard Species}
\usage{
standard_species(df)
}
\arguments{
\item{df}{The data.frame in which the isotope ratios
for all species are stored.}
}
\value{
A data.frame with the same number of rows as the original data frame,
   with the standardized English species name and other standardized grouping
   information for each entry
}
\description{
Takes a dataframe and identifies the species column (either "species", "Species",
"taxon" or "Taxon") and matches the listed species against a dictionary list of
standardized English names in the "species_lookup_table" The standardized names
are then looked up in a standardized plotting table (plot_lookup_table) to
find the plotting parameters.
}
\examples{
x <- standard_species(test_df)
}

